- move main logic to the lib.rs to make it test-able - de-compose main into parsing and execution
